♻️ Extract product page pricing, staking, share, and book list pieces
Pricing selector and mobile sticky bar become components that emit
intent; social share and book-list status move to composables; the
staking accordion body becomes ProductStakingInfo. Purchase handlers
stay in the page untouched.
